local-mmap: fix -Wsign-compare
__builtin_popcountl() returns int and nb_blocks is unsigned which means
-Wsign-compare. In theory, it should be an harmless warning but let's
still shut it down and cast nb_blocks (which should never be < 0) to
int.

+++ local-mmap.c
@@ -113,7 +113,7 @@ local_create_mmap_block(struct iio_buffer_pdata *pdata,
 		goto out_free_priv;
 	}
 
-	if (__builtin_popcountl(ppdata->mmap_block_mask) == ppdata->nb_blocks) {
+	if (__builtin_popcountl(ppdata->mmap_block_mask) == (int) ppdata->nb_blocks) {
 		/* All our allocated blocks are used; we need to create one more. */
 		priv->idx = ppdata->nb_blocks;
 		req.id = 0;
